Use single printnode_api namespace
