+++ printnode_api/__init__.py
@@ -1,8 +1,13 @@
-"""Preferred import namespace for the community-maintained PrintNode client.
-
-The historical import package is ``printnodeapi``. It remains supported for
-compatibility, while new code may import from ``printnode_api`` to match the
-maintained distribution name more closely.
-"""
-
-from printnodeapi import *  # noqa: F401,F403
+from .gateway import Gateway
+from .auth import (
+    ApiError,
+    ClientError,
+    Unauthorized,
+    TooManyRequests,
+    ServerError,
+    NetworkError,
+    TimeoutError,
+    TooManyRedirectsError,
+    ConnectionError,
+    HttpError,
+    RequestError)
