--- printnodeapi/__init__.py
@@ -1,13 +1,0 @@
-from .gateway import Gateway
-from .auth import (
-    ApiError,
-    ClientError,
-    Unauthorized,
-    TooManyRequests,
-    ServerError,
-    NetworkError,
-    TimeoutError,
-    TooManyRedirectsError,
-    ConnectionError,
-    HttpError,
-    RequestError)
