+++ pyproject.toml
@@ -51,7 +51,7 @@ dev = [
 ]
 
 [tool.hatch.build.targets.wheel]
-packages = ["printnodeapi", "printnode_api"]
+packages = ["printnode_api"]
 
 [tool.pytest.ini_options]
 testpaths = ["tests"]
